Move stateLayer Tailwind plugin to its own file
